backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
scikit-learn
python-multipart
PyPDF2
python-docx
requests
beautifulsoup4
google-api-python-client
google-auth
